+++ crates/prover-services/sharp-service/src/lib.rs
@@ -1,6 +1,7 @@
 pub mod client;
 pub mod config;
 pub mod error;
+mod types;
 
 use std::str::FromStr;
 
@@ -33,12 +34,8 @@ impl ProverClient for SharpProverService {
                 let fact_info = get_fact_info(&cairo_pie, None)?;
                 let encoded_pie =
                     snos::sharp::pie::encode_pie_mem(cairo_pie).map_err(ProverClientError::PieEncoding)?;
-                let res = self.sharp_client.add_job(&encoded_pie).await?;
-                if let Some(job_key) = res.cairo_job_key {
-                    Ok(combine_task_id(&job_key, &fact_info.fact))
-                } else {
-                    Err(ProverClientError::TaskInvalid(res.error_message.unwrap_or_default()))
-                }
+                let (_, job_key) = self.sharp_client.add_job(&encoded_pie).await?;
+                Ok(combine_task_id(&job_key, &fact_info.fact))
             }
         }
     }
@@ -47,6 +44,9 @@ impl ProverClient for SharpProverService {
         let (job_key, fact) = split_task_id(task_id)?;
         let res = self.sharp_client.get_job_status(&job_key).await?;
         match res.status {
+            // TODO : We would need to remove the FAILED, UNKNOWN, NOT_CREATED status as it is not in the sharp client response specs :
+            // https://docs.google.com/document/d/1-9ggQoYmjqAtLBGNNR2Z5eLreBmlckGYjbVl0khtpU0
+            // We are waiting for the official public API spec before making changes
             CairoJobStatus::FAILED => Ok(TaskStatus::Failed(res.error_log.unwrap_or_default())),
             CairoJobStatus::INVALID => {
                 Ok(TaskStatus::Failed(format!("Task is invalid: {:?}", res.invalid_reason.unwrap_or_default())))
@@ -59,7 +59,7 @@ impl ProverClient for SharpProverService {
                 if self.fact_checker.is_valid(&fact).await? {
                     Ok(TaskStatus::Succeeded)
                 } else {
-                    Ok(TaskStatus::Failed(format!("Fact {} is not valid or not registed", hex::encode(fact))))
+                    Ok(TaskStatus::Failed(format!("Fact {} is not valid or not registered", hex::encode(fact))))
                 }
             }
         }
@@ -74,6 +74,13 @@ impl SharpProverService {
     pub fn with_settings(settings: &impl SettingsProvider) -> Self {
         let sharp_cfg: SharpConfig = settings.get_settings(SHARP_SETTINGS_NAME).unwrap();
         let sharp_client = SharpClient::new(sharp_cfg.service_url);
+        let fact_checker = FactChecker::new(sharp_cfg.rpc_node_url, sharp_cfg.verifier_address);
+        Self::new(sharp_client, fact_checker)
+    }
+
+    pub fn with_test_settings(settings: &impl SettingsProvider, port: u16) -> Self {
+        let sharp_cfg: SharpConfig = settings.get_settings(SHARP_SETTINGS_NAME).unwrap();
+        let sharp_client = SharpClient::new(format!("http://127.0.0.1:{}", port).parse().unwrap());
         let fact_checker = FactChecker::new(sharp_cfg.rpc_node_url, sharp_cfg.verifier_address);
         Self::new(sharp_client, fact_checker)
     }